+++ app.js
@@ -29,6 +29,13 @@ class AppBootHook {
     // 可以用来加载应用自定义的文件，启动自定义的服务
 
     // 例如：创建自定义应用的示例
+    // Configure default routes
+    const { router } = this.app;
+
+    // Framework default routes
+    router.post('/upload-pre', this.app.middleware.preUpload());
+    router.post('/upload-transfer', this.app.middleware.transfer());
+
   }
 
   async willReady() {