+++ app/middleware/auth.js
@@ -3,11 +3,8 @@ module.exports = (options, app) => {
   // 基于jwt实现的注册用户的token鉴权
   return async (ctx, next) => {
 
-    // 排除不需要验证 token 的路由
-    if ((options.allowed.indexOf(ctx.request.url) > -1) || (options.allowed.indexOf('*') > -1)) return await next(options);
-
     // 获取token
-    const tokenKey = app.config.private.TokenFieldName; // 从配置中读取token的键名
+    const tokenKey = (app.config.private && app.config.private.TokenFieldName) ? app.config.private.TokenFieldName : 'cbegg-token'; // 从配置中读取token的键名
     const token = ctx.header[tokenKey];
     if (!token) ctx.authFailed({ msg: '您没有足够权限访问接口!', code: 401 });
 
@@ -22,7 +19,6 @@ module.exports = (options, app) => {
     }
 
     // 把 user 信息挂载到全局ctx上
-    user.user_id = user.id; // 解决修改user_id后需要全盘修改的问题
     user.token = token;
     ctx.auth = { user };
     await next(options);