+++ app/router.js
@@ -6,10 +6,10 @@
 module.exports = app => {
   const { router, controller } = app;
   router.get('/', controller.home.index);
-  // 上传预检
-  router.post('/upload-pre', app.middleware.preUpload());
-  // 上传
-  router.post('/upload-transfer', app.middleware.transfer());
+  // // 上传预检
+  // router.post('/upload-pre', app.middleware.preUpload());
+  // // 上传
+  // router.post('/upload-transfer', app.middleware.transfer());
 
 
 };