+++ config/config.default.js
@@ -3,8 +3,9 @@
 'use strict';
 
 const path = require('path');
-const os = require('os');
-const { port } = require('../libs/sequelize-db/db.config');
+const HOST = '127.0.0.1';
+const PORT = 7001;
+const UPLOAD_DIR = './app/public/upload';
 
 /**
  * @param {Egg.EggAppInfo} appInfo app info
@@ -16,92 +17,64 @@ module.exports = appInfo => {
    **/
   const config = exports = {};
 
-  // use for cookie sign key, should change to your own and keep security
-  config.keys = appInfo.name + '_1682650998331_7872';
-  // logger
+  config.keys = appInfo.name + '_0964776560948769_6748758';
+
   config.logger = {
     level: 'OFF', // log level
   };
 
-  // 自定义部分
-  const userConfig = {};
-
-  // 配置暴露的IP
-  userConfig.cluster = {
+  config.cluster = {
     listen: {
-      hostname: '127.0.0.1',
-      port: 7001,
+      hostname: HOST,
+      port: PORT,
     },
   };
 
   // 跨域配置
-  userConfig.security = {
+  config.security = {
     csrf: {
       enable: false,
     },
     domainWhiteList: [ '*' ],
   };
-  userConfig.cors = {
-    origin: '*',
-    // origin: 'http://127.0.0.1:7001',
+  config.cors = {
+    origin: '*', // 允许所有跨域
     credentials: true, // 允许Cook可以跨域
     allowMethods: 'GET,HEAD,PUT,POST,DELETE,PATCH,OPTIONS',
   };
 
   // 上传配置
-  userConfig.upload = {
+  config.upload = {
     tokenExpiresIn: 10, // 上传token有效期 (秒)
   };
-  userConfig.multipart = {
-    fileSize: '10mb', // 上传文件大小限制
+  config.multipart = {
+    fileSize: '2mb', // 上传文件大小限制
     mode: 'stream',
   };
 
-  // 中间件配置
-  // userConfig.middleware = [ 'exception', 'auth' ];
-  // config.exception = {
-  //   // 配置项，应用可覆盖
-  //   match: '/',
-  //   env: appInfo.env,
-  // };
-  // // 不需要鉴权的路由
-  // userConfig.auth = {
-  //   allowed: [
-  //     '/',
-  //     '*',
-  //   ],
-  // };
-  // 鉴权jwt设置
-  userConfig.jwt = {
+  // jwt设置
+  config.jwt = {
     secret: 'commonBackendEggjs', // 密钥
-    expiresIn: 60 * 60 * 24 * 1000, // 有效期（秒）
+    expiresIn: 60 * 60 * 24, // 有效期（秒）
   };
 
-  // redis配置
-  // userConfig.redis = {
-  //   client: require('../libs/redis/redis.config'),
-  // };
-
   // orm设置
-  userConfig.sequelize = require('../libs/sequelize-db/db.config');
+  config.sequelize = require('../libs/sequelize-db/db.config');
 
   // 静态化配置   访问路径如：http://127.0.0.1:7001/static/images/logo.png
-  userConfig.static = {
-    domain: `http://127.0.0.1:7001`,
+  config.static = {
+    domain: `http://${HOST}:${PORT}`,
     prefix: '/static/',
-    dir: path.join(appInfo.baseDir, './app/public/upload'), // 这里是当前项目目录下的public/upload
+    dir: path.join(appInfo.baseDir, UPLOAD_DIR), // 上传文件本地路径
     dynamic: true, // 如果当前访问的静态资源没有缓存，则缓存静态文件，和`preload`配合使用；
     preload: false,
     maxAge: 31536000, // in prod env, 0 in other envs
     buffer: true, // in prod env, false in other envs
   };
 
-
-  // 隐私配置，不push的部分
-  userConfig.private = require('./private.js');
+  // config.private = require('../private/config.js');
 
   return {
     ...config,
-    ...userConfig,
   };
 };