+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.0.5] - 2024-12-27
+
+### Fixed
+- Fixed upload pre-request and tansfer
 
 ## [1.0.4] - 2024-12-26